requirements.txt
torch==2.3.1+cu124
torchvision==0.18.1+cu124
torchaudio==2.3.1+cu124
albumentations==1.3.1
streamlit
opencv-python
numpy
Pillow
tqdm
scikit-learn